+++ account-api/src/main/java/com/momsitter/api/ui/AccountController.java
@@ -1,6 +1,6 @@
 package com.momsitter.api.ui;
 
-import com.momsitter.authiorization.ui.webconfig.AccountAuthenticationPrinciple;
+import com.momsitter.api.ui.webconfig.AccountAuthenticationPrinciple;
 import com.momsitter.domain.Account;
 import com.momsitter.service.AccountService;
 import com.momsitter.service.dto.AccountInfoResponse;
