+++ account-api/src/main/java/com/momsitter/api/ui/webconfig/AuthInterceptor.java
@@ -1,6 +1,6 @@
-package com.momsitter.authiorization.ui.webconfig;
+package com.momsitter.api.ui.webconfig;
 
-import com.momsitter.authiorization.infrastructure.AuthorizationExtractor;
+import com.momsitter.api.infrastructure.AuthorizationExtractor;
 import com.momsitter.authiorization.infrastructure.JwtTokenProvider;
 import com.momsitter.common.exception.UnauthorizedException;
 import org.springframework.http.HttpMethod;
