+++ authorization-jwt/src/main/java/com/momsitter/authiorization/service/AuthService.java
@@ -3,8 +3,8 @@ package com.momsitter.authiorization.service;
 import com.momsitter.authiorization.exception.InvalidAuthTokenException;
 import com.momsitter.authiorization.exception.LoginFailException;
 import com.momsitter.authiorization.infrastructure.JwtTokenProvider;
-import com.momsitter.authiorization.ui.dto.auth.TokenRequest;
-import com.momsitter.authiorization.ui.dto.auth.TokenResponse;
+import com.momsitter.authiorization.service.dto.TokenRequest;
+import com.momsitter.authiorization.service.dto.TokenResponse;
 import com.momsitter.domain.Account;
 import com.momsitter.domain.AccountId;
 import com.momsitter.domain.repository.AccountRepository;
