+++ authorization-jwt/src/main/java/com/momsitter/authiorization/service/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package com.momsitter.authiorization.ui.dto.auth;
+package com.momsitter.authiorization.service.dto;
 
 public class TokenRequest {
